all check and ready

+++ BackendShopNow/src/main/java/com/ShopNow/DAO/cartDao.java
@@ -67,13 +67,13 @@ public class cartDao {
 
 
     }
-    public Integer deleteUser(shoppingCart cartData){
+    public Integer deleteProduct(shoppingCart cartData){
         String userId = cartData.getUserId();
         String productId = cartData.getProductId();
 
         try{
-            String deleteCartItem = "delete from shoppingCart where productId=?";
-            Integer update = this.cartJdbc.update(deleteCartItem, productId) ;
+            String deleteCartItem = "delete from shoppingCart where productId=? and userId=?";
+            Integer update = this.cartJdbc.update(deleteCartItem, productId, userId) ;
             return update;
         }catch (Exception e){
             System.out.println(e.getMessage());
